requirements.txt
absl-py==2.2.0
accelerate==1.5.2
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
distro==1.9.0
dlib==19.24.6
easyocr==1.7.2
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
fire==0.7.0
flash_attn==2.7.4.post1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
# genai==2.1.0  # 已移除：该 PyPI 包依赖 openai<0.28，与 openai>=1.x 冲突；项目使用 google-genai，代码未使用本包
google-auth==2.39.0
google-genai==1.12.1
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llava @ git+https://github.com/LLaVA-VL/LLaVA-NeXT.git@09e5840d5589ad2d6a8656c0a60f21ae134b3309
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==2.1.2
mpmath==1.3.0
multidict==6.4.3
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.2
openai==1.77.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
proglog==0.1.10
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9.1
safetensors==0.5.3
scenedetect==0.6.6
scikit-image==0.25.2
scipy==1.15.2
sentencepiece==0.2.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
supervision==0.25.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tifffile==2025.3.13
tiktoken==0.3.3
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
# torch==2.6.0
# torchaudio==2.6.0
# torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yapf==0.43.0
yarl==1.20.0
